chore: move `animated-three-d-text-prisms` to components folder, adjust import path, and update prism height styling
